decouple core rpc types from cfxcore types
