+++ crates/client/src/rpc/helpers/mod.rs
@@ -2,8 +2,10 @@
 // Conflux is free software and distributed under GNU General Public License.
 // See http://www.gnu.org/licenses/
 
+mod block_provider;
 mod subscribers;
 
+pub use block_provider::{build_block, build_header};
 pub use cfx_rpc::helpers::{
     poll_filter::{
         limit_logs, PollFilter, SyncPollFilter, MAX_BLOCK_HISTORY_SIZE,
